fixed Main manifest for jar file and King making rochade multiple times/after his first move. Added new move function to figure, which is called with the new position of the figure.
