+++ src/main/java/functional/Board.java
@@ -1,4 +1,4 @@
-package util;
+package functional;
 
 import functional.figure.Figure;
 import functional.figure.PlayerDependentFigure;
@@ -170,6 +170,7 @@ public class Board {
 		if (getFigure(to) != null) getFigure(to).kill(this);
 		set(to, getFigure(from));
 		set(from, null);
+		getFigure(to).move(to);
 		moves.add(new DoubleSet<>(from, to));
 	}
