+++ src/main/java/functional/figure/Figure.java
@@ -1,6 +1,6 @@
 package functional.figure;
 
-import util.Board;
+import functional.Board;
 import util.Player;
 import util.Position;
 
@@ -55,4 +55,6 @@ public abstract class Figure {
 	public void actionOnRoundEnd(Board board, Position from) {}
 
 	public void kill(Board board) {}
+
+	public void move(Position from) {}
 }
