+++ src/main/java/gui/Canvas.java
@@ -1,6 +1,6 @@
 package gui;
 
-import util.Board;
+import functional.Board;
 import functional.figure.Figure;
 import functional.figure.figures.*;
 import functional.figure.special.KillLinked;
